feat(actions.tsx,history.tsx,history-card.tsx,sidebar.tsx,dynamic-category-icon.tsx): updated components to use dynamic category icon. refactored sidebard cards to own component
